#474409 [C#テキスト]17章演習問題
Practice17-1-1において、LineCounterの機能を切り分けてクラスを作成しました。
