+++ 章末問題17/Practice17-1-1/LineCounter/LineCounter/LineCounter.csproj
@@ -50,6 +50,7 @@
     <Compile Include="LineCounterProcessor.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ToHankakuNumber.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
